+++ src/components/MetadataViewProvider.ts
@@ -241,8 +241,15 @@ export class MetadataViewProvider
   private async getBaseNodeChildren(
     element: MetadataNode
   ): Promise<MetadataNode[]> {
+    const baseName = element.baseNodeName;
+
+    /**
+     * @todo: REFACTOR: Use OCAPI system_object_definition_search call to filter
+     *    results for only system or custom object definitions on the server
+     *    before returning results.
+     */
     const _callSetup: ICallSetup = await this.service.getCallSetup(
-      element.baseNodeName,
+      baseName,
       'getAll',
       {
         count: 200,
@@ -250,6 +257,7 @@ export class MetadataViewProvider
       }
     );
 
+    // Call the OCAPI service.
     const _callResult = await this.service.makeCall(_callSetup);
 
     // If the API call returns data create a tree.
@@ -257,16 +265,24 @@ export class MetadataViewProvider
       // Add the display name to the custom objects so that they can be
       // easily identified as custom.
       return _callResult.data.filter(obj => {
-        return obj.object_type !== 'CustomObject';
-      }).map(sysObj => {
-        let name = sysObj.object_type;
+        return baseName === 'systemObjectDefinitions' ?
+          (obj.object_type !== 'CustomObject') :
+          (obj.object_type === 'CustomObject' && obj.display_name);
+      }).map(filterdObj => {
+        // Get the display name for the tree node.
+        let name = '';
+        if (baseName === 'systemObjectDefinitions') {
+          name = filterdObj.object_type;
+        } else if (baseName === 'customObjectDefinitions') {
+          name = filterdObj.display_name.default;
+        }
 
         // Create a MetaDataNode instance which implements the TreeItem
         // interface and holds the data of the document type that it
         // represents.
         return new MetadataNode(name, TreeItemCollapsibleState.Collapsed, {
           parentId: 'root.systemObjectDefinitions',
-          objectTypeDefinition: new ObjectTypeDefinition(sysObj),
+          objectTypeDefinition: new ObjectTypeDefinition(filterdObj),
           displayDescription: ' '
         });
       });
@@ -274,7 +290,7 @@ export class MetadataViewProvider
   }
 
   /**
-   * Gets the children the root element.
+   * Gets the base nodes of the tree that can be expanded for viewing data types.
    * @param {MetadataNode} element - The MetadataNode instance.
    * @return {Promise<MetadataNode[]>}
    */
@@ -290,8 +306,14 @@ export class MetadataViewProvider
     );
 
     // Get the VSCode settings for display of each base tree node.
+    // - Show System Object Definitions
     const showSystemObjects: boolean = Boolean(
       workspaceConfig.get('explorer.systemobjects')
+    );
+
+    // - Show Custom Object Definitions
+    const showCustomObjects: boolean = Boolean(
+      workspaceConfig.get('explorer.customobjects')
     );
 
     // If the user config is enabled, then show the option.
@@ -303,6 +325,20 @@ export class MetadataViewProvider
           {
             parentId: 'root',
             baseNodeName: 'systemObjectDefinitions'
+          }
+        )
+      );
+    }
+
+    // If display of Custom Object Definitions is enabled, add node to tree.
+    if (showCustomObjects) {
+      metaNodes.push(
+        new MetadataNode(
+          'Custom Object Definitions',
+          TreeItemCollapsibleState.Collapsed,
+          {
+            parentId: 'root',
+            baseNodeName: 'customObjectDefinitions'
           }
         )
       );